requirements.txt
alembic==1.7.7
asttokens==2.0.5
backcall==0.2.0
black==22.3.0
click==8.1.2
decorator==5.1.1
executing==0.8.3
Flask==2.1.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
ipdb==0.13.9
ipython==8.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
Pygments==2.11.2
pyparsing==3.0.8
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.35
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
traitlets==5.1.1
typing-extensions==4.1.1
wcwidth==0.2.5
Werkzeug==2.1.1
zipp==3.8.0
